requirements.txt
django
djangorestframework
django-cors-headers
resend
whitenoise
django_resized
django-ckeditor-5

